Added presentation of weather and forecast + favoriting system
I've added two simple functions that will be further improved on that show the responses for current weather and for the forecast on screen. I've also added the system for saving favorites and displaying current weather for them when you come back to the page. It currently works on city selection, it should be moved to a button next to the city name during the polishing section